Move access list parser, add tests
